Update HubCallout.tsx
A developer-centric platform and share Where you can discover, deploy and interact with any Polywrapper in the ecosystem.
We are paving the road to endless collaboration and curation possibilities. Test and Integrate web3 protocols quickly on the web with our GraphQL Playground, and publish your packages to decentralised hosting such as IPFS. Soon you'll be able to explore an endless ocean of wrappers, by querying tags like `multisig`, `defi`, or `vesting`. A more semantic web3 that's easy to compose!